addressed review comments from karsten
